[server] Change file size dt to Long

+++ server/src/main/kotlin/ru/andrey/remote/controller/request/DeviceFilesInfoRequest.kt
@@ -14,4 +14,4 @@ data class DeviceFilesInfoRequest(
         val filesInfo: List<FileInfoRequest>
 )
 
-data class FileInfoRequest(val path: String, val size: Int?)
+data class FileInfoRequest(val path: String, val size: Long?)
